Refactored use of content terminology as attributes where appropriate
Necessary to improve readability by referring to attributes as the data structure being manipulated rather than content (which was too generic). In some cases, due to line width, I used `pairs` instead of `attributes`. The goal of all this is to clarify -- to the reader -- that you are dealing with a hash.

Milestone: patch