Added time transfomer
Necessary to provide the time at which the configuration was loaded. This is handy for situations where you might want date/time calculations based off of when the configuration was loaded.

This was extracted from the link:https://alchemists.io/projects/rubysmith[Rubysmith] project but is also used in the following projects:

* link:https://alchemists.io/projects/milestoner[Milestoner]
* link:https://alchemists.io/projects/gemsmith[Gemsmith]
* link:https://alchemists.io/projects/hanamismith[Hanamismith]

Milestone: minor