Updated registry to find symbolized loaders and transformers
Necessary to leverage newly added functionality that allows you to load an Etcher specific loader and/or transformer by symbol.

The additional `.find` method provides lightweight functionality for unwrapping the monad result for logging purposes while still allowing the logger to be injected. This might need to be extracted to an object in the future.

Milestone: minor