Fixed step-based failure details
Necessary to ensure the loaders and transformers surface all failures. The bulk of this work is with the loaders since they were only logging debugging information previously. Now the loaders ensure critical errors are captured. The `Builder` was updated to ensure both the loaders and transformers are processed in the same manner so any failure causes all other operations to be skipped.

The failure payload has been updated to provide more details:

* `step`: Identifies the step in which the command is run (i.e. load, transform, validate, or model).
* `constant`: Identifies the command that failed.
* `payload`: The formatted error message or hash if there is more details to provide.

The resolver will soon be updated to make use of these changes.

Finally, Reek was updated to ignore having too many statements because keeping all of the exception logic together is better than attempting to split this apart further.

Milestone: minor