Added format transformer retainers
Necessary to deal with situations where you need to partially format a string but don't want to remove specifiers that need to be resolved until later. This is similar in behavior to the original `pass_throughs` behavior but reduces typing because the specifier and value are identical.

This means you can now specify retainers (new) and mappings (originally known as pass throughs) which are then combined together as pass throughs for final formatting. The goal is the same for both where the former is for identical specifiers and the latter is for specifiers with different values.

Milestone: minor